requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==0.14.0
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==1.35.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpcio==1.41.0rc2
idna==3.2; python_version >= '3'
markdown==3.3.4; python_version >= '3.6'
numpy==1.21.2
oauthlib==3.1.1; python_version >= '3.6'
opencv-python==4.5.3.56
pandas==1.3.3
pillow==8.3.2
protobuf==3.18.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.6.0
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
wheel==0.37.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
